pyproject.toml
[project]
name = "bodo-os"
version = "0.1.0"
description = ""
authors = [
    {name = "Kolja Sam",email = "github@koljapluemer.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
